Refactor typing, DAE IO, and shader nodes
Broad refactor adding stronger typing and small API cleanups across the addon. Highlights: introduce XmlReader and improve COLLADA parsing/writing (Accessor, DaeMatrix.flatten_matrices, animation flattening), move/clean numpy imports, and make DAE reader/writer more type-safe. Add NodeSignature abstractions and a BeamDisplayInject node; refactor NodeTree/NodeGroup builder types and many Blender operator/presets typing fixes. Fixes in PackedVector (generic typing, to_numpy_array) and equality, parser/builder mesh handling, msgpack typing, and several pyright/type compatibility annotations and minor bug fixes.

+++ src/beamng/cdae/io/dae_asset.py
@@ -14,8 +14,8 @@ class Imposter:
     dimension: int = 256
     equatorSteps: int = 16
     includePoles: bool = False
-    polarAngle: int = 0
-    polarSteps: int = 0,
+    polarAngle: float = 0
+    polarSteps: int = 0
     size: int = 25
 
     def apply_detail(self, detail: CdaeV31.Detail):
